backend/agents/browser_agent/requirements.txt
browser-use==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0 